add roaring dependency, separate equality delete code
