+++ CMakeLists.txt
@@ -25,6 +25,7 @@ set(EXTENSION_SOURCES
     src/base_manifest_reader.cpp
     src/manifest_list_reader.cpp
     src/manifest_file_reader.cpp
+    src/deletes/equality_delete.cpp
     src/metadata/iceberg_transform.cpp
     src/metadata/iceberg_table_schema.cpp
     src/metadata/iceberg_partition_spec.cpp
@@ -73,6 +74,9 @@ if(NOT CMAKE_SYSTEM_NAME STREQUAL "Emscripten")
     include_directories(${CURL_INCLUDE_DIRS})
 endif()
 
+# Roaring is installed via vcpkg and used here
+find_package(roaring CONFIG REQUIRED)
+
 # Reset the TARGET_NAME, the AWS find_package build could bleed into our build -
 # overriding `TARGET_NAME`
 set(TARGET_NAME iceberg)
@@ -91,6 +95,10 @@ if(NOT CMAKE_SYSTEM_NAME STREQUAL "Emscripten")
     target_link_libraries(${EXTENSION_NAME} PUBLIC ${CURL_LIBRARIES})
     target_link_libraries(${TARGET_NAME}_loadable_extension ${CURL_LIBRARIES})
 endif()
+
+# Link Roaring library
+target_link_libraries(${EXTENSION_NAME} PUBLIC roaring::roaring roaring::roaring-headers roaring::roaring-headers-cpp)
+target_link_libraries(${TARGET_NAME}_loadable_extension roaring::roaring roaring::roaring-headers roaring::roaring-headers-cpp)
 
 install(
   TARGETS ${EXTENSION_NAME} ${TARGET_NAME}_loadable_extension
